1.19.4-pre1

+++ src/main/java/net/minecraft/world/item/enchantment/FrostWalkerEnchantment.java
@@ -39,10 +39,10 @@ public class FrostWalkerEnchantment extends Enchantment {
     public static void onEntityMoved(LivingEntity param0, Level param1, BlockPos param2, int param3) {
         if (param0.isOnGround()) {
             BlockState var0 = Blocks.FROSTED_ICE.defaultBlockState();
-            float var1 = (float)Math.min(16, 2 + param3);
+            int var1 = Math.min(16, 2 + param3);
             BlockPos.MutableBlockPos var2 = new BlockPos.MutableBlockPos();
 
-            for(BlockPos var3 : BlockPos.betweenClosed(param2.offset((double)(-var1), -1.0, (double)(-var1)), param2.offset((double)var1, -1.0, (double)var1))) {
+            for(BlockPos var3 : BlockPos.betweenClosed(param2.offset(-var1, -1, -var1), param2.offset(var1, -1, var1))) {
                 if (var3.closerToCenterThan(param0.position(), (double)var1)) {
                     var2.set(var3.getX(), var3.getY() + 1, var3.getZ());
                     BlockState var4 = param1.getBlockState(var2);
